Added a full option to support persisting all changes to the proxied Model.
By default sequelize-temporal persist only changes, and saves the previous state in the history table.
I wanted a way to persist every state that landed in the main table at some point. This includes intial creation and latest state.
This allows to only query the hostory table to get the full history of an entity, and also integrates nicely with sequelize paranoid mode.